Split semantic checks
