+++ include/semantic.h
@@ -6,7 +6,12 @@
 #include "symtable.h"
 
 
-/* Type check helpers */
+/* Error handling */
+void semantic_set_error(size_t line, size_t column);
+void semantic_print_error(const char *msg);
+
+/* Expression/statement checking helpers */
+int eval_const_expr(expr_t *expr, int *out);
 type_kind_t check_expr(expr_t *expr, symtable_t *vars, symtable_t *funcs,
                        ir_builder_t *ir, ir_value_t *out);
 int check_stmt(stmt_t *stmt, symtable_t *vars, symtable_t *funcs,
@@ -15,8 +20,5 @@ int check_stmt(stmt_t *stmt, symtable_t *vars, symtable_t *funcs,
 int check_func(func_t *func, symtable_t *funcs, symtable_t *globals,
                ir_builder_t *ir);
 int check_global(stmt_t *decl, symtable_t *globals, ir_builder_t *ir);
-
-/* Print the last semantic error with source location */
-void semantic_print_error(const char *msg);
 
 #endif /* VC_SEMANTIC_H */
